requirements.txt
-i https://pypi.org/simple
blinker==1.6.2; python_version >= '3.7'
cachelib==0.10.2; python_version >= '3.7'
click==8.1.6; python_version >= '3.7'
dnspython==2.4.1; python_version >= '3.8' and python_version < '4.0'
flask==2.3.2
flask-session==0.5.0
gunicorn==21.2.0
importlib-metadata==6.8.0; python_version < '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
numpy==1.25.2
packaging==23.1; python_version >= '3.7'
pandas==2.0.3
pymongo==4.4.1
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tzdata==2023.3; python_version >= '2'
werkzeug==2.3.6; python_version >= '3.8'
zipp==3.16.2; python_version >= '3.8'
